Abstract ts parsing into libts.
